+++ src/index.ts
@@ -45,6 +45,7 @@ import type { ReminderRecurrenceRule } from './schemas/interfaces/reminder-recur
 import type { RemindersList } from './schemas/interfaces/reminders-list';
 import type { RequestPermissionOptions } from './schemas/interfaces/request-permission-options';
 import type { SelectCalendarsWithPromptOptions } from './schemas/interfaces/select-calendars-with-prompt-options';
+import type { SelectCalendarsWithPromptResult } from './schemas/interfaces/select-calendars-with-prompt-result';
 import type { UpdateRemindersListOptions } from './schemas/interfaces/update-reminders-list-options';
 import type { UpdateRemindersListResult } from './schemas/interfaces/update-reminders-list-result';
 import type { EventEditAction } from './schemas/types/event-edit-action';
@@ -97,6 +98,7 @@ export type {
   RequestAllPermissionsResult,
   RequestPermissionOptions,
   SelectCalendarsWithPromptOptions,
+  SelectCalendarsWithPromptResult,
   UpdateRemindersListOptions,
   UpdateRemindersListResult,
 };